Edit: Navbar fixture - now it is only in one file
